Clean up + de-warnify my code
Messed up a bunch of stuff in UIConstrainable. Eventually, when you make a constraint, it will return a reference that you can actually modify and have it reflect the changes in the constraint solver immediately. For now, just return a reference to a copy.

Also anything that returned multiple constraints I just changed to return void for now.